scanner/requirements.txt
kafka-python>=2.0.2
bleak>=0.22.0
Pillow>=9.0.0
